#277 [refactor] SingleEventHandler로 네이밍 변경 및 로직 변경
