+++ designsystem/src/main/java/hous/release/designsystem/util/single_event/SingleEventHandleStrategy.kt
@@ -1,0 +1,5 @@
+package hous.release.designsystem.util.single_event
+
+fun interface SingleEventHandleStrategy {
+    fun handle(event: () -> Unit)
+}
