+++ designsystem/src/main/java/hous/release/designsystem/util/single_event/SingleEventHandler.kt
@@ -1,0 +1,9 @@
+package hous.release.designsystem.util.single_event
+
+class SingleEventHandler(
+    private val singleEventHandleStrategy: SingleEventHandleStrategy = ThrottledDurationStrategy()
+) {
+    fun handle(event: () -> Unit) {
+        singleEventHandleStrategy.handle(event)
+    }
+}
